+++ semcore/input-mask/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [5.15.1] - 2023-12-14
+
+### Added
+
+- ``
+
+### Fixed
+
+- Text cursor position after keypress `Backspace`.
+
 ## [5.15.0] - 2023-12-06
 
 ### Changed